Use the Spring Initializr (and Boot 2.2)
I rewrote the document and the code (mostly changing package names) to use the Spring Initializr. I also modified the build files to use Spring Boot 2.2.0, which necessitated replacing the Gradle wrappers. Note that there are no tests (other than the one I pulled in from the src directory that the Initializr created).